physics-ai/requirements.txt
# requirements.txt - Agentic Physics System
streamlit>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.1.0
chromadb>=0.4.0
numpy>=1.21.0
scipy>=1.7.0
sympy>=1.12
matplotlib>=3.4.0
plotly>=5.3.0
opencv-python>=4.9.0
pybullet>=3.2.5
pybullet-data>=1.0.4
gymnasium>=0.29.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=8.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=1.3.0
scikit-learn>=1.3.0
fastapi>=0.104.0
celery>=5.3.0
redis>=5.0.0
pydantic>=2.0.0
networkx>=3.0